1 try
Tic tac toe is a simple and popular game that can be played by two players on a 3x3 grid. The goal is to make a line of three Xs or Os, either horizontally, vertically, or diagonally, before the opponent does. The game is also known as noughts and crosses or Xs and Os